Add AiTutorView component and refactor AiTutorFloatingActionButton package
